js: Add forward-compatibility mechanism with require.js
The main motivation for these big changes is forward-compatibility: the idea is
that with the new 'apiVersion' property, I can break backwards compatibility by
loading a different JS plugin proxy at runtime. Poking around, looking at async
JS loaders, I found require.js. The only drawback was that I would have to redo
most of the existing code to work with it. So... this is that.