Replace the manifest approach with a more direct approach.
In order to increase security, the Shell will download directly from a URL
it creates itself to help prevent XSS injection or other errors.

The new URL, /download-extension/, is passed a set of GET parameters, some
of which are unused right now. One, server_uuid, is something we can pass
to the plugin which is guaranteed a round trip unmodified.