Allow more than one review on a version.
With this change, we start remodeling the role of a review to be more like a
commen on a piece of code, designed to facilitate conversation, like Splinter
and Review Board do, rather than a canonical once-given approve/reject.

One of these days I should probably just embed Review Board inside SweetTooth.
It's a Django app...