Make errorreports Extension specific, not ExtensionVersion specific
The guarantee that a user would grab the correct version to report an error
to is very slim... and with our upcoming multiversion feature, we can't even
guarantee that "the latest version" is the one the user installed. We're
better off detecting the extension version and sticking it in the pre-built
error report.