Add new UI so that extension authors can see and manipulate their versions
With the new multiversion feature, "what version is the user getting?"
doesn't have a straightforward answer. Add a new UI to the bottom of extension
details panes that gives extension authors the ability to see which version
of an extension the user is getting based on their shell version, and the
ability to activate or deactivate extension versions.