review: Allow the option to select the version to diff against
With the new multiversion feature, diffing against the previous version
isn't too useful -- an extension author may be creating multiple versions
side by side to target different shell versions. We don't have a UI for
this, but we should allow reviewers the opportunity to diff the changes
against any version.