Add and from now on populate 'popularity': (enables - disables)
An extension's popularity is the number of enables minus the number of
disables. Because Django's ORM isn't quite up to the task of doing this
math in the sorting of the query quite yet, we track this as three
separate fields in the database: enables, disables and popularity.

From now on, the JavaScript associated with enabling/disabling an
extension will also ping back to the server to keep track of this
anonymous data, used entirely for statistics.