Move from "ratings" to "discussions"
See http://youtube-global.blogspot.com/2009/09/five-stars-dominate-ratings.html
for some discussion of the same problem, talking about YouTube.

Star ratings as a guideline aren't quite useful. The ratings feature was being
used as a discussion board on the extension's page, rather than a critique of
an extension. To solve this, follow the YouTube strategy, following a binary
like/dislike system. The implementation of this system will come later, so
I'm not gutting the ratings code -- the data migration for it still needs to
exist.