Move diffview entirely client-side
Instead of calculating the diffview HTML and sending it over
to the client, construct the DOM elements on the client with
some jQuery. This will give us better line number support,
as well as make things generally line up in the diff.

For now, disable highlighting. I have a plan for adding back
highlighting, but it will be done entirely client-side. The
main problem with the current approach is that any syntax
differences will show up in the diff viewer, resulting in
things that look like this:

    tx">foo bar

as part of an inline change region. It's because something
changed syntactic categories, so we see that in the diff
output.