Make ratings optional
"Ratings" were meant to be in-depth reviews of an extension, explaining what a
user liked or didn't like about an extension, which should hopefully give the
extension author insightful feedback about the quality of what they created and
the path forward for developing and improving their extension.

That never happened. The "ratings" section is used like a discussions board
for the extension. The section isn't really designed as a discussion board,
without proper replies or pagination or anything modern (go Django!).
Part of this may be that there wasn't a separate discussion board for users
to get their gripes about GNOME Shell and "I LOVE IT"s out.

Rather than introduce a new system, let's try and jiggle the UI around a bit.
The comment form is now divided into three choices: you can "Leave a comment",
a "rating", or a "bug report" (which links to the error report system as
before).

Comments are just specialized versions of ratings with the "rating" field
fixed to -1. The UI to pick a star rating doesn't appear, and the list of
comments won't show the stars at all.