upload: Rework the upload code
First of all, use transactions. Second, refactor and split out the big
mess that is parse_metadata_json, so we don't have to do any extraneous
saving. Thirdly, fix and add new tests so that we're sure that this code
is working. Fourthly, remove the dedicated "Upload a new version" page,
as this just clutters the code up. Looking up an existing version by
UUID should be more than enough.