extensions: Fix base shell-version list visibility
If someone queries for a non-existant "3.2.5" version, we need to query the
base version "3.2". Since we had a nice object-oriented API to grab the base
version, this wouldn't work, since we had no object for "3.2.5" to query the
base version off of. Fall back to explicit querying, and remove the OOP API
so we don't make that mistake again.