extensions: Fix complex visibility
If we had an extension which had a visible version and a invisible
version, it was possible that properties of the invisible version
would be exposed on a query list. Use a subquery to solve this. Example:

 Extension Foo
   Visible Version 1
     Shell version 3.2
   Invisible Version 2
     Shell version 3.3.90

Query for shell_version=3.3.90 would show Extension Foo, as it was
a visible extension that matched.