review: Exclude botched uploads when trying to grab diffs
In some cases, extension authors can "corrupt" an extension by uploading
something that causes the server to 500 out. In this case, an admin usually
has to delete the extension manually and make the author re-upload. While
we should fix the cause of the 500, let's try and not "corrupt" the extension
by looking for something good to diff against.