upload: Fix upload page
Obviously, the mistake that I made here was really silly, modifying
the DB and not committing. I didn't catch it because transactions
are monkey patched out during testcases because why not, Django!

So I was sitting here thinking that my test cases were actually
testing the code, when they were not. Turns out you need to use
a TransactionTestCase, which is completely obvious. Thanks, Django!