js: Import mustache, introduce a new templates system
If we want to sanely translate the website, we need to have some
basic form of templating. Having hesitated, waiting for the jQuery
team to develop a templating framework for the fifth time, I think
I've waited long enough. Mustache it is.

Port some basic things over to the new templates system, as a test.