review: Automatically approve harmless extension versions
When a new extension version is uploaded, we diff it against the last
approved version -- if the changes to its files are "safe", we automatically
approve the new version.

Currently, safe changes are defined by filenames and file types -- any
updates to images, translation files, or metadata.json is considered safe,
and more may be added to the list in the future.

A similar feature is planned that checks for future 'safe' versions
upon approval of other versions of the same extension.