extensions: Remove the extra "Submit" step in between uploading and reviewing
It was visibly confusing users, for almost no gain, especially now that users
can reject their own extensions. We're still keeping STATUS_NEW around, as it
is required for backwards compatibility and all that. At some point, we may
reject all STATUS_NEW versions of active extensions, and email the authors
of extensions that never got past the dumb submit UI wall.

While we're at it, let's remove the "step" numbers, and a lot of the
instructions. The UI might need some more design in the future to be a bit
more intuitive and featureful, but for now let's assume that it's good enough.