css: Merge sweettooth.css and template.css
These two have been competing with each other for a little while.
The original intention was that template.css was the template for
any gnome-branded website, and sweettooth.css contained the
sweettooth-specific styles. As time went on, this separation became
less and less clear, and I started modifying the template.css file.

Just merge the two, now.